Refactor to use context instead of shader for general stuff
