+++ shaders/simple_f.glsl
@@ -3,11 +3,11 @@
 
 out vec4 FragColor;
 
-in vec3 vertexColor;
+in vec4 vertexColor;
 // uniform vec3 color;
 
 void main()
 {
   // FragColor = vec4(color * vertexColor, 1.0);
-  FragColor = vec4(vertexColor, 1.0);
+  FragColor = vertexColor;
 }
